MINOR: tinfo: add a dynamic thread-group context
The thread group info is not sufficient to represent a thread group's
current state as it's read-only. We also need something comparable to
the thread context to represent the aggregate state of the threads in
that group. This patch introduces ha_tgroup_ctx[] and tg_ctx for this.
There's also a tg_ctx pointer in the thread so that it can be reached
from other threads. It's indexed on the group id and must be cache-line
aligned. For now the entry is empty.