MINOR: epoll: delete unused updates on close()
In epoll we know we won't need to do anything on an FD after a close(),
and tests show that __fd_clo() is called roughly half of the time with
the last FD from the local update list, which possibly makes sense if a
close has to appear after a polling change resulting from an incomplete
read or the end of a send().

We can detect this and remove the last entry, which gives less work to
do during the update() call, and eliminates most of the poll_drop_fd
event reports.

Note that while tempting, this must not be backported because it's only
safe to be done now that fd_delete_orphan() clears the update mask as
we need to be certain not to miss it:
  - if the update mask is kept up with no entry, we can miss future
    updates ;
  - if the update mask is cleared too fast, it may result in failure
    to add a shared event.