MINOR: fd: add two functions fd_take_tgid() and fd_drop_tgid()
These ones will be used to try to grab a reference to the desired
thread group in the FD. These will be used during the time the
running_mask is not yet set, to make sure we atomically get a
running mask that matches the tgid. Once the running_mask is set,
the tgid reference can be dropped.