SERVER: introduce new server setting `forward-proxy`
This commit adds a new SERVER setting named `forward-proxy`. When
a server is marked as such, haproxy will issue HTTP CONNECT negotiation
on every connection opened with the server. If the CONNECT returns a
HTTP 200 (OK) response, the response is cleared from the TCP
connection. Otherwise, upstream connection handshake is considered
failed, which closes both upstream and downstream TCP connections.

`forward-proxy` has only one (mandatory) parameter, which is an haproxy
session variable name which stores the hostname of the server that an
incoming connection tries to reach. The variable should be created
in the frontend configuration by inspecting incoming data. See [1] for
an example configuration.

Note that only session variables are supported. So, the `sess.` prefix
in variable name is mandatory. See [2] for documentation of variable
types. If during server connection establishment the variable is not
found in session variables, upstream connection fails and downstream
connection is rejected. Thus, this setting should be used only if
content inspection is used in the frontend.

[1]

The following example delays an incoming downstream connection up to 5
seconds (to wait for stream data). If incoming connection is TLS, the
SNI is stored in variable named `sess.sni_name`, and the connection is
directed to `ds` backend. The `ds` backend uses the variable
`sess.sni_name` to fetch the hostname and send it in HTTP CONNECT
message, as the server is marked as forward proxy.

frontend tlsf
	bind :443
	tcp-request inspect-delay 5s
	tcp-request content set-var(sess.sni_name) req.ssl_sni if { req.ssl_hello_type 1 }
	tcp-request accept if { req.ssl_hello_type 1 }
	use_backend ds if { req.ssl_hello_type 1 }

backend ds
	server my-server my-server.local:3128 forward-proxy sess.sni_name

[2] https://www.haproxy.com/documentation/hapee/latest/onepage/#http-request%20set-var