LISTENER: allow setting SO_MARK on listening socket
Setting SO_MARK is available as a TCP action in various connection states.
However, these connection states are all after the TCP negotiation had
completed.

However, MARK may be used as a zone indicator, meaning that iptables
can assign packets by zone and only have lookups done in that zone.
Therefore it relies on this value and as such TCP negotiation cannot be completed
without setting the correct mark for the socket.

This commit adds the ability to set the MARK value on the listening socket
so that TCP any TCP connection would automatically inherit this value.

The setting is done as part of the 'bind' directive, e.g.:

bind :443 transparent interface tungre1 mark 0x8